requirements.txt
Flask
flask-cors
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
numpy
pillow
pytesseract
opencv-python-headless